# backend/internal/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Lofi Radio API
  version: "1.0.0"
  description: >
    Backend API for the Lofi Radio player. Handles song library, playback queue,
    player state, and audio streaming. Real-time updates are delivered via WebSocket at `/ws`.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/queue:
    get:
      summary: List queue
      operationId: listQueue
      tags: [queue]
      responses:
        "200":
          description: Current playback queue
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/QueueItem"
                  total:
                    type: integer
    post:
      summary: Add song to queue
      operationId: addToQueue
      tags: [queue]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [song_id]
              properties:
                song_id:
                  type: string
                  format: uuid
      responses:
        "201":
          description: Queue item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueItem"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/queue:play-next:
    post:
      summary: Play song immediately
      description: >
        Inserts the requested song at the front of the queue and broadcasts `skip_now`
        over WebSocket so connected audio clients interrupt their current track and
        start playing the requested song.
      operationId: playNext
      tags: [queue]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [song_id]
              properties:
                song_id:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Song queued for immediate play
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: song queued for immediate play
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Song:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        duration_secs:
          type: number
          format: float
        size_bytes:
          type: integer
          format: int64
        added_at:
          type: string
          format: date-time
    QueueItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        song_id:
          type: string
          format: uuid
        song:
          $ref: "#/components/schemas/Song"
        position:
          type: integer
        source:
          type: string
          enum: [auto, manual]
        added_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string